test: split trace compare variant coverage
